third_party/blink/web_tests/wpt_internal/js-self-profiling/foreign-script-cors-disallow.sub.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>

  <script src="resources/profile-utils.js"></script>

  <script id="foreign-script" src="http://{{hosts[][www1]}}:{{ports[http][0]}}/wpt_internal/js-self-profiling/resources/external-script.js"></script>
</head>
<body>
  <script>
    promise_test(async t => {
      assert_not_equals(window.externalScriptFunction, undefined);
      const trace = await ProfileUtils.profileFunction(externalScriptFunction);

      assert_false(ProfileUtils.containsFrame(trace, {
        name: 'externalScriptFunction',
      }), 'foreign function without CORS is omitted');

      const foreignScriptUrl = document.querySelector('#foreign-script').src;
      assert_false(ProfileUtils.containsFrame(trace, {
        resourceId: trace.resources.indexOf(foreignScriptUrl),
      }), 'no frame exists with non-CORS resource');

      assert_false(ProfileUtils.containsResource(trace, foreignScriptUrl),
                   'foreign resource is omitted');

    }, 'function from foreign-origin script without CORS is not included in trace');
  </script>
</body>
</html>
